lighter_hedger.py
"""
Lighter Exchange Hedge Module
Handles hedging positions on Lighter when StandX orders are filled
"""
import logging
import time
from typing import Optional, Dict, Any
from decimal import Decimal

logger = logging.getLogger(__name__)


class LighterHedger:
    """
    Lighter exchange hedging client.

    Note: This is a template implementation. You'll need to integrate
    with the actual Lighter API based on their documentation.
    """

    def __init__(self, config):
        """
        Initialize Lighter hedger.

        Args:
            config: Config instance
        """
        self.config = config
        self.enabled = config.get("exchanges.lighter.enabled", True)
        self.api_url = config.get("exchanges.lighter.api_url", "https://api.lighter.xyz")

        # Get credentials from environment
        self.private_key = config.get_lighter_private_key()

        # Position tracking
        self.current_position = 0.0  # Net position on Lighter

        logger.info(f"Initialized Lighter Hedger (enabled={self.enabled})")

    def connect(self) -> bool:
        """
        Connect and authenticate with Lighter exchange.

        Returns:
            True if successful
        """
        if not self.enabled:
            logger.info("Lighter hedging is disabled")
            return False

        try:
            # TODO: Implement Lighter authentication
            # This depends on Lighter's API specification
            # Example:
            # self.session = LighterSession(api_key=self.api_key, secret=self.api_secret)
            # self.session.authenticate()

            logger.info("Connected to Lighter exchange")
            return True

        except Exception as e:
            logger.error(f"Failed to connect to Lighter: {e}")
            return False

    def place_hedge_order(self, side: str, quantity: float, price: Optional[float] = None) -> bool:
        """
        Place a hedge order on Lighter.

        Args:
            side: "buy" or "sell" (opposite of StandX fill)
            quantity: Order quantity
            price: Limit price (None for market order)

        Returns:
            True if successful
        """
        if not self.enabled:
            logger.warning("Lighter hedging disabled, skipping hedge order")
            return False

        try:
            logger.info(f"Placing Lighter hedge: {side} {quantity} @ {price or 'market'}")

            # TODO: Implement actual Lighter order placement
            # Example API call structure:
            # order_type = "limit" if price else "market"
            # response = self.session.create_order(
            #     symbol=self.config.get("trading.symbol"),
            #     side=side,
            #     order_type=order_type,
            #     quantity=str(quantity),
            #     price=str(price) if price else None
            # )

            # Simulate successful order for now
            logger.info(f"Lighter hedge order placed successfully")

            # Update position tracking
            qty_signed = quantity if side == "buy" else -quantity
            self.current_position += qty_signed

            return True

        except Exception as e:
            logger.error(f"Failed to place Lighter hedge order: {e}")
            return False

    def get_position(self, symbol: str = None) -> float:
        """
        Get current position on Lighter.

        Args:
            symbol: Trading symbol (uses config default if None)

        Returns:
            Position size (positive for long, negative for short)
        """
        if not self.enabled:
            return 0.0

        try:
            # TODO: Implement actual position query
            # Example:
            # symbol = symbol or self.config.get("trading.symbol")
            # positions = self.session.get_positions()
            # for pos in positions:
            #     if pos['symbol'] == symbol:
            #         return float(pos['quantity'])
            # return 0.0

            # Return tracked position for now
            return self.current_position

        except Exception as e:
            logger.error(f"Failed to get Lighter position: {e}")
            return 0.0

    def close_position(self, symbol: str = None) -> bool:
        """
        Close entire position on Lighter.

        Args:
            symbol: Trading symbol

        Returns:
            True if successful
        """
        if not self.enabled:
            return False

        try:
            position = self.get_position(symbol)
            if abs(position) < 0.0001:
                logger.info("No position to close on Lighter")
                return True

            # Determine side to close
            side = "sell" if position > 0 else "buy"
            quantity = abs(position)

            logger.info(f"Closing Lighter position: {side} {quantity}")

            # Place closing order
            success = self.place_hedge_order(side, quantity)

            if success:
                self.current_position = 0.0
                logger.info("Lighter position closed successfully")

            return success

        except Exception as e:
            logger.error(f"Failed to close Lighter position: {e}")
            return False

    def sync_position(self):
        """Sync position tracking with actual exchange position"""
        if not self.enabled:
            return

        try:
            # TODO: Query actual position from Lighter
            # actual_position = self.get_position()
            # self.current_position = actual_position
            pass

        except Exception as e:
            logger.error(f"Failed to sync Lighter position: {e}")

    def get_balance(self) -> Dict[str, float]:
        """
        Get account balance on Lighter.

        Returns:
            Dictionary with balance information
        """
        if not self.enabled:
            return {"balance": 0.0, "available": 0.0}

        try:
            # TODO: Implement actual balance query
            # Example:
            # balance_data = self.session.get_balance()
            # return {
            #     "balance": float(balance_data.get("total", 0)),
            #     "available": float(balance_data.get("available", 0))
            # }

            return {"balance": 0.0, "available": 0.0}

        except Exception as e:
            logger.error(f"Failed to get Lighter balance: {e}")
            return {"balance": 0.0, "available": 0.0}


# Example integration notes:
"""
To integrate with actual Lighter API:

1. Find Lighter's Python SDK or API documentation
2. Install required packages (e.g., pip install lighter-python-sdk)
3. Update the connect() method with proper authentication
4. Implement place_hedge_order() with actual API calls
5. Implement get_position() to query real positions
6. Implement get_balance() to query account balance

Common patterns for crypto exchange APIs:
- REST API for orders, positions, balances
- WebSocket for real-time updates
- API key + secret for authentication
- HMAC signatures for request signing

Example Lighter integration (hypothetical):
```python
from lighter import LighterClient

class LighterHedger:
    def __init__(self, config):
        self.client = LighterClient(
            api_key=config.get_env("LIGHTER_API_KEY"),
            secret_key=config.get_env("LIGHTER_API_SECRET")
        )

    def place_hedge_order(self, side, quantity, price=None):
        order = self.client.place_order(
            symbol="BTC-USDT",
            side=side,
            type="market" if price is None else "limit",
            quantity=quantity,
            price=price
        )
        return order.success
```
"""
